adding view file
